backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beanie>=2.0.1",
    "certifi>=2025.11.12",
    "fastapi>=0.123.9",
    "google-auth>=2.43.0",
    "httpx>=0.28.1",
    "motor>=3.7.1",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
